secp256k1: Make field value set int take uint16.
This modifies the SetInt method on a field value to take a uint16
instead of a uint in order to make it impossible to misuse it since it
has a precondition of a max value of 2^26 - 1.  In practice, the code
currently only ever calls it with 0 and 1, but in order to prepare for
exporting the type, it's nicer to prevent the potential misuse.

+++ dcrec/secp256k1/field.go
@@ -171,7 +171,7 @@ func (f *fieldVal) Set(val *fieldVal) *fieldVal {
 //
 // The field value is returned to support chaining.  This enables syntax such
 // as f := new(fieldVal).SetInt(2).Mul(f2) so that f = 2 * f2.
-func (f *fieldVal) SetInt(ui uint) *fieldVal {
+func (f *fieldVal) SetInt(ui uint16) *fieldVal {
 	f.Zero()
 	f.n[0] = uint32(ui)
 	return f
